requirements.txt
Adafruit-DHT==1.3.0
RPi.GPIO==0.6.2
paho-mqtt==1.2
